refactor: separate filter to FilterForm component
